Strings, interpolation, percent escapes, and percent arrays
For percent-escaped strings that use the curly bracket but don't use
interpolation (like `%q{foo}`) I had to also include a subscope for
interpolation. If you try something like this:

    %q{no #{interpolation} allowed}

it will try to end the scope after the first `}`. The subscope detects
interpolation and tokenizes it as more percent-escaped string.

For percent arrays, a recursive include is used so that the
tokenization doesn't end early in a situation like this:

    %w(some %w(words) here)

This captures the whole line, instead of missing ` here)`