Methods
To capture method with and without parenthesized arguments, we need two
separate rules.

Othwerise, methods that have one line of body, and don't have any
arguments (or really, no parenthesis) won't show the code-folding
indicators. The reason is that methods with no parenthesis fail to
parse the method name on a single line for some reason. They end on the
next line with the first bit of non parenthesis word characters.

This makes the method body appear to have no lines if it's a one-line
method. Nova (probably) just hides the folding indicators if it's
there's no content in the method itself.

I couldn't figure out why it doesn't stop on the line-break. My guess
is it's has something to do with the anchored subscopes, but even
removing them didn't solve it.