Add formatting with crystal tool format
Things I learned:

The crystal formatter will only write to STDOUT when using the `-`
argument, and then it only will read from STDIN not a file path.
https://github.com/crystal-lang/crystal/issues/1863#issuecomment-1532305
42

To write to STDIN, a WritableStream must be set up and then written to
when it's ready.

STDOUT writes one line at a time, I think. Or at least not all at once.
Because of this, the code needs to append it to some buffer if there's
more than one line to get the whole output (like the contents of a
file).

To keep the cursor position in the file, use `editor.edit(...)` with a
callback and `edit.replace(...)` inside the callback.
https://docs.nova.app/api-reference/text-editor/#editcallback-options

It's probably better for portability to use `configWorkspace` for
extension settings than `config`. The latter is App wide so if you're
doing something like setting a path executable that could change per
project, it'd be a real pain to keep changing that every time you
switched projects.

You can early return on a bunch of things like if the language for the
document is _not_ `crystal` or some setting isn't set. Not really
learned per se, but a nice idea to keep these processes responsive.