Fix formatting after save
In my excitement to ship v1.1, I somehow did not notice that when
Format on Save was enabled, it would save, _then_ format which makes
the document unsaved.

To fix this, I wrapped my formatting code in a Promise that resolves
when it writes the file and rejects when there's an error.

https://docs.nova.app/api-reference/text-editor/#onwillsavecallback-this
value