Refactor formatter
This moves the formatter into its own file and converts to a
Promise-based API. This allows for a really nice separation of
concerns. One function handles the actual formatting of the code and
knows nothing about the document or Nova APIs. A separate function
deals with the document stuff like collecting the initial text and
interacting with Nova.